+++ js/src/pages/lesson/ContentLesson.js
@@ -42,7 +42,7 @@ const ContentLesson = ({ lesson, nextLesson, prevLesson }) => {
 
           return (
             <Route path={`/section-${index + 1}`} key={index} exact>
-              <Box mt={[2, 2, 0]}>
+              <Box>
                 <LessonGrid>
                   <Box display="flex" justifyContent="space-between" alignItems="center">
                     <Hidden smDown>